Divide protobuf unit/integration tests
Coming changes to these tests will require integration tests to use a
protobuf_test package to avoid an import cycle with generated code.
